update policy to use DSL and added demo security package on shorten url app

+++ orthos/core/src/main/java/dev/igordesouza/orthos/core/signal/SignalResult.kt
@@ -13,6 +13,6 @@ data class SignalResult(
     val signalId: SignalId,
     val signalType: SignalType,
     val triggered: Boolean,
-    val confidence: Float,
+    val confidence: Double,
     val metadata: Map<String, String> = emptyMap()
 )
